requirements.txt
# Patchy - Multi-Agent GitHub Security Bot
# Core Framework
Flask==3.1.0
python-dotenv==1.1.0

# GitHub Integration
PyGithub==2.6.0
requests==2.32.3

# NVIDIA AI Endpoints (via OpenAI SDK)
# Model: stepfun-ai/step-3.5-flash (40 RPM)
openai>=1.0.0

# Background Jobs (for Railway worker)
redis==5.0.1
rq==1.15.1

# File Locking (for scans.json concurrent access)
filelock>=3.0.0

# Security Scanner
semgrep>=1.60.0

# Session Management
Flask-Session==0.8.0

# Development
gunicorn==23.0.0
